Moved `generator` and created exectuable path
