+++ exe/caze
@@ -1,0 +1,3 @@
+ #!/usr/bin/env ruby
+
+require 'caze/generator'
